requirements.txt
fastapi
uvicorn
requests
groq
langchain-groq
langchain_core
python-dotenv
streamlit
